requirements.txt
grpcio==1.65.1
grpcio-tools==1.54.0
tensorflow==2.17.0
opencv-python
